New --with-ftgl212 to declare FTGL 2.1.2 location
Let --with-ftgl declare location of FTGL version > 2.1.2,
for instance ftgl-2.1.3-rc5 which is the latest relased
version to date.

This will define a new HAVE_FTGL_NEWER212 macro, which is
used in OpenGl_FontMgr.hxx and OpenGl_FontMgr.cxx.

Let configure abort if Freetype or FTGL are not found and
HAVE_GL is set to yes, compilation will then fail anyway.
We could disable X11, but this is not what user expects.

Source fixes needed to compile with recent FTGL have been
submitted upstream, see
   http://www.opencascade.org/org/forum/thread_20128/