Fix Tk and X11 dependencies in TKDraw and TKViewerTest
Modify src/Draw/Draw_Window.cxx to include tk.h on
Windows only if _TK is defined.  All Tk instructions
are protected by _TK on Windows, so this change is safe.
On Unix machines, tk.h is included unconditionnally.

As all other source files compile without Tk on
Windows, it is very likely that Tk support on Windows
is not well tested, so it does not hurt if it is
disabled.  We can easily reenable it later.

Modify TKViewerTest in a similar manner, Tk is used
only on Unix, and X11 includes and libraries are added
since they are used in src/ViewerTest/ViewerTest_ViewerCommands.cxx