Rename OpenCASCADE and OPENCASCADE into ${PROJECT_NAME} in all CMakeLists.txt
Our project is not endorsed by OpenCASCADE, make sure to not
confuse users.  As our project is OCE, users will see this name,
but it will automatically be changed back to OPENCASCADE if this
build system is adopted by OpenCASCADE by modifying a single line.

Remove copyright notice, this is redundant since all files are
released under the OpenCASCADE Public License, and this notice
may confuse users.