Remove PACKAGE, VERSION, and RETSIGTYPE macros
PACKAGE and VERSION are unused. Remove them.
RETSIGTYPE is not used within OCE. Also, it is obsolete ( https://lists.isc.org/pipermail/inn-workers/2009-April/016669.html ), so it shouldn't affect any system headers that OCE uses.

STDC_HEADERS is not used in OCE, but it does show up in some system headers - so I left it.