Change OSD_FontMgr.cxx so that TKernel does not depend on X11
This class uses X11 calls to get the list of available fonts,
which is then used by OpenGl_FontMgr.cxx; see
  http://www.opencascade.org/org/forum/thread_21351/
for details about why this is a bad idea.

Here we use an environment variable to get the list of directories
where fonts have to be looked for.  If it is unset, default paths
are set when compiling:
 * On Darwin, /usr/X11/lib/X11/fonts:/Library/Fonts:/System/Library/Fonts
 * On other Unix, /usr/share/X11/fonts:/usr/share/fonts/X11

Closes issue #12.

Tkernel indirectly depends on libdl through libX11, so this commit
must be applied after ca78eea, otherwise applications may have
unresolved symbols when linking against TKernel.