Do not redefine memmove as an alias for memcpy
In math_Memory.hxx, memmove is redefined as an alias for
memcpy.  This is plain wrong, it may break user code if
memmove is used because memory regions do overlap.

The reverse_move function is kept to not break API, but
it is useless and should be removed too.

Fixes issue #94.