Let users specify COMPONENTS in find_package(OCE)
Those components are OpenCascade toolkits, or in other words,
library names.
If REQUIRED is also specified, a FATAL_ERROR is thrown if at least
one library is missing.  If REQUIRED is not specified, processing
continues even if a library is missing.  User can check afterwards
if OCE_<toolkit>_FOUND is set to true or false.