Workaround GCC (< 4.4) bugs in gp_Mat and gp_XYZ
Report and patch provided by Roman Lygin at
  http://www.opencascade.org/org/forum/thread_19603/

For instance, Standard_ConstructionError is raised
in gp_Trsf::SetValues, the reason is that gp_Mat::Subtract
does not work as expected (only the first row is
subtracted).

It is not clear what makes GCC buggy, but compilers should
be smart enough to compute multiple array indices when
compiling, so this change should not introduce any performance
problems.  It would be better to go one step further and
remove macros, but it will make this patch much longer.