Bugfix: Points swapped in BRepExtrema_DistanceSS
Fix issue #260.  Thanks to AkitaMan for the report and patch.
Distance between edge/vertex and face/vertex was broken because
points added into SeqSolShape1 and SeqSolShape2 were swapped.

Distance between face/edge was also broken, but for a slightly
different reason, the line
  if (TRI_SOLUTION(SeqSolShape1,Pt1) || TRI_SOLUTION(SeqSolShape2,Pt2))
was wrong because Pt1 is on shape 2 and Pt2 is on shape 1.  Other
lines were right.

Rewrite code so that index 1 (resp. 2) always refer to shape 1 (resp. 2).