Link against FreeImage (instead of FreeImagePlus) on non-Windows
FreeImage support is available in TKService and TKOpenGl.
In TKService, only the C API is used, so we can link
against freeimage.  This is better because on some platforms
(Fedora for instance), symbols of freeimage are not duplicated
into freeimageplus, which causes unresolved symbols when linking
with --as-needed.

In TKOpenGl, the FreeImage support was not enabled, copy
it from TKService.  It is used only in OpenGl_Workspace::Print,
and this method works only on Windows.

Remove #pragma from src/OpenGl/OpenGl_Workspace_2.cxx.