Take CSF_GraphicShr environment variable into account
The handling of libTKOpenGl has been fully redesigned
in OCCT 6.6.0.  The CSF_GraphicShr environment variable
is now useless, it does nothing.

We added OCE_DEFAULT_CSF_GraphicShr a while back so that
libTKOpenGl is found at runtime without having to set
LD_LIBRARY_PATH.

But the drawback is that if there is a libTKOpenGl located
at OCE_DEFAULT_CSF_GraphicShr, it is not possible to
load a library located at another location.

Reinstate the usage of the CSF_GraphicShr environment in
order to allow overriding OCE_DEFAULT_CSF_GraphicShr.