Made OCE_BUILD_TYPE initialize to CMAKE_BUILD_TYPE if available.
Prior to this patch, CMAKE_BUILD_TYPE was overwritten to be OCE_BUILD_TYPE
even if the user specified it on command line. This patch allows the
user to configure CMAKE_BUILD_TYPE in the normal way.

This change is especially important for incorporation into other projects
where CMAKE_BUILD_TYPE usually controls the build type for all
sub-projects.