Reverted removal 'get_target_property' to discover ARCHIVE_OUTPUT_DIRECTORY.
If 'CMAKE_ARCHIVE_OUTPUT_DIRECTORY' is used to customize the output
folders of oce-win-bundle libraries, the oce build will set
FREETYPE_LIBRARY and related variables to the wrong paths.

This patch fixes that bug.

The commit that inserted this removal in the first place
(https://github.com/tpaviot/oce/pull/465) did not provide any explination
for it.