[mingw-warning-fix][maybe-uninitialized]
If Abs(Coord - Arr(i)) == Tol here in Hunt(), then it's possible for the variable that Iloc references to be uninitialized without raising an error. Hunt() is used in Locate2Coord() defined later on in the file, and it's possible for the variable N there to be used in comparisons without being initialized.