Fix stack overflow in meshSimplePolygon
This is a fairly ugly hack that replaces a recursive call to
meshSimplePolygon with a manually-managed stack in the special cases
where the input polygon is split right at the beginning or right at the
end. A better fix would probably refactor the method more
comprehensively - this change tries to keep the original structure of
the code as much as possible.